pyproject.toml
[project]
name = "skills-mcp-server"
version = "0.1.0"
description = "MCP server exposing skills as tools and resources"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
skills-mcp-server = "server:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
